CT_Grid3D désormais indéxé sur des size_t au lieu des int.
git-svn-id: http://rdinnovation.onf.fr/svn/plugin-lvox@36 f1561d03-89af-ab4a-a97c-2bf9dc1f41eb

+++ pluginlvox/tools/lvox_computebeforethread.cpp
@@ -45,7 +45,7 @@ void LVOX_ComputeBeforeThread::run()
     CT_Beam beam(NULL, NULL);
     QVector3D origin, direction;
 
-    int progressStep = n_points / 20;
+    size_t progressStep = n_points / 20;
 
     for (size_t i = 0 ; i < n_points; i++)
     {
